RSR-10 Show a change history for issues
Every change to issues and related objects is saved in versions of this
object. Versions are saved with the request id and the id of the member
who changed the issue.

After each issue change request a background job is started to create a
history object. A history groups all changes of one request for to show
a change set for all related objects to the user.

The change history is shown to members if they are related to this
issue. All changes to a project can also be seen on the project page to
easily see recent activities.