RSR-11 Issue can not assign children as a parent
If a issue assigns a child as a parent a infinite loop is created. The
issues is now validating that the parent is not included in the
children.

The validation is not thread save. If the parent is saved as a child
between validation and issue update the issue will have a child as a
parent.